requirements.txt


fastapi-cors
